Set up CircleCI using Ruby 2.4.10
A few changes were needed:

- Ignore config/database.yml (which may vary across different environments)

- Change the way we require models to make sure that it doesn't fail

Sometimes this was failing because it was trying to create a Tag (Sequel model) when
the relation didn't exist. This happens when you are setting up your environment for
the first time.

That is why I had to require models only when necessary and I removed the require 'app'.

- Make sure we require the models before we use them

This is not great (it is a data migration) but it is a workaround to avoid problems
when trying to run migrations when the database is not created yet.

- Upgrade sequel_pg, pg, and sequel. Drop thin and start using puma.

This is an attempt to pay off tech debt and use more recent versions of these gems.

- Remove Gemfile.lock from repository

This will be better for CircleCI and testing with different rubies and versions of bundler.